Rename task.h into task_base.h
